+++ src/storage/BlockBuilder.cpp
@@ -18,12 +18,12 @@ void BlockBuilder::add(const std::string &key, const std::string &value) {
         counter_ = 0;
     }
 
-    uint32_t key_size = key.size();
-    uint32_t value_size = value.size();
+    uint64_t key_size = key.size();
+    uint64_t value_size = value.size();
 
-    buffer_.append((char*)&key_size, sizeof(uint32_t));
+    buffer_.append((char*)&key_size, sizeof(uint64_t));
     buffer_.append(key);
-    buffer_.append((char*)&value_size, sizeof(uint32_t));
+    buffer_.append((char*)&value_size, sizeof(uint64_t));
     buffer_.append(value);
 
     last_key_ = key;
@@ -36,16 +36,16 @@ bool BlockBuilder::empty() const {
 }
 
 std::string BlockBuilder::finish() {
-    uint32_t restarts_postion = buffer_.size();
-    uint32_t num = restarts_.size();
+    uint64_t restarts_postion = buffer_.size();
+    uint64_t num = restarts_.size();
     // append size
-    buffer_.append((char*)&num, sizeof(uint32_t));
+    buffer_.append((char*)&num, sizeof(uint64_t));
     // append restarts_
     for(auto i : restarts_){
-        buffer_.append((char *)&i, sizeof(uint32_t));
+        buffer_.append((char *)&i, sizeof(uint64_t));
     }
     // append postion
-    buffer_.append((char *)&restarts_postion, sizeof(uint32_t));
+    buffer_.append((char *)&restarts_postion, sizeof(uint64_t));
     // finished
     finished_ = true;
 
@@ -64,6 +64,6 @@ void BlockBuilder::reset() {
 
 size_t BlockBuilder::estimated_size() const {
     return (buffer_.size() + 
-            restarts_.size() * sizeof(uint32_t) + 
-            sizeof(uint32_t) * 2);
+            restarts_.size() * sizeof(uint64_t) + 
+            sizeof(uint64_t) * 2);
 }
